Add base CardMCell
It will be used as a base cell for other M-sized card cells, so we can
have the common logic and layout in a single place.
