+++ decidim-core/app/cells/decidim/card_m/comments_counter.erb
@@ -1,0 +1,3 @@
+<%= icon("comment-square", class: "icon--small") %>
+<%= " " %>
+<%= comments_count %>
