+++ decidim-core/app/cells/decidim/card_m/header.erb
