+++ decidim-core/app/cells/decidim/card_m/show.erb
@@ -1,19 +1,18 @@
 <div class="column" id="<%= dom_id(model) %>">
-  <article class="card card--meeting">
+  <article class="card card--<%= dom_class(model) %>">
     <div class="card__content">
       <%= render :header %>
 
       <div class="card__text">
         <div class="card__text--paragraph">
           <%= description %>
-          <%= render partial: "decidim/shared/tags.html", locals: { resource: model, tags_class_extra: "tags--meeting" } %>
+          <%= render :tags %>
         </div>
       </div>
     </div>
 
-    <%= render :date %>
-    <%= render :address %>
+    <%= render :data %>
     <%= render :status %>
-    <%= render :footer if model.can_be_joined_by?(current_user) %>
+    <%= render :footer if shows_footer? %>
   </article>
 </div>
