+++ decidim-meetings/app/cells/decidim/meetings/meeting_m/data.erb
@@ -1,0 +1,2 @@
+<%= render :date %>
+<%= render :address %>
