+++ decidim-meetings/app/cells/decidim/meetings/meeting_m/footer.erb
@@ -1,5 +1,7 @@
-<div class="card__footer">
-  <div class="card__support">
-    <%= cell "decidim/meetings/join_meeting_button", model %>
+<% if shows_footer? %>
+  <div class="card__footer">
+    <div class="card__support">
+      <%= cell "decidim/meetings/join_meeting_button", model %>
+    </div>
   </div>
-</div>
+<% end %>
