--- decidim-meetings/app/cells/decidim/meetings/meeting_m/status.erb
@@ -1,18 +1,0 @@
-<div class="card__status">
-  <ul class="card-data">
-    <li class="card-data__item">
-      <%= l(model.created_at.to_date, format: :decidim_short) %>
-    </li>
-    <li class="card-data__item">
-      <%= cell "decidim/follow_button", model, inline: true %>
-    </li>
-    <li class="card-data__item">
-      <%= link_to resource_path do %>
-        <%= with_tooltip t("decidim.meetings.models.meeting.fields.comments") do %>
-          <%= icon("comment-square", class: "icon--small") %>
-          <%= model.comments.count %>
-        <% end %>
-      <% end %>
-    </li>
-  </ul>
-</div>
