+++ decidim-meetings/app/cells/decidim/meetings/meeting_m/tags.erb
@@ -1,0 +1,1 @@
+<%= render partial: "decidim/shared/tags.html", locals: { resource: model, tags_class_extra: "tags--meeting" } %>
