+++ decidim-meetings/lib/decidim/meetings/view_model.rb
@@ -3,11 +3,6 @@
 module Decidim
   module Meetings
     class ViewModel < Decidim::ViewModel
-      include TranslatableAttributes
-      include LayoutHelper
-      include Decidim::Meetings::MeetingsHelper
-      include Decidim::SanitizeHelper
-      include Decidim::Meetings::Engine.routes.url_helpers
     end
   end
 end
